HTML/CSS: Fix selection issue
If you hammer on the left-click mouse button like I do you will run into
issues where anchor links that are acting as buttons and triggering
Javascript are causing issues for you.  However, there is a whole host
of reasons we don't want to use the <button> tag (primarily, styling),
so we add a couple of CSS helper classes to effect the right behavior in
our desktop browser.

[ Documentation: None ]
[ Testing: Deployed locally and mashed clicker ]