nft-creator: Update slot before submitting transaction
There is an edge case (which I was guilty of doing often during QA),
where a user updates the timestamp before updating their blockfrost key
or connecting the wallet and no slot gets generated.  This change fixes
this by confirming the slot update before submitting the mint txn.  If
there is nothing returned, we assume that some sort of error happens
since we typically return a promise (because of the async call to
blockfrost parameters this operation requires).

Note that we add a little bit of security around our calls to
e.preventDefault() for all the functions in this file based on our new
call to a function with e === undefined.

[ Documentation: Nothing ]
[ Testing: Manual QA and messing around ]