Secrets: Remove the main/test split for secrets.js
This was really confusing the interface for blockfrost and introducing
behaviors that would occur in that class.  This change streamlines
MAIN/TEST_*_ADDR/PROJ into just PAYMENT_ADDR and BLOCKFROST_PROJ.  This
has some knock-on effects that we need to debug manually in nft-creator,
but the main focus is drop-payment.

[ Documentation: README.md ]
[ Testing: npm run build, manual QA ]