Vending Machine: Add support for generating the keys in your browser
Version 1 of this required you to input your vending machine keys from
what you generated in your command line.  This proved to be a
prohibitive barrier for several folks, so we provide a button here to
generate the keys in your browser.

Note that this requires us to use an "Enterprise Address" because we
don't intend the address to ever be used for staking, just for
temporarily collecting payment.  It also simplifies the code so we don't
need to generate staking keys.

[ Documentation: None except for UI updates ]
[ Testing: Manually locally, then deployed to Squarespace ]