Mint Machine: Support auto-detect updates for NFTs
For Cardano NFTs, "updating" token values is represented as performing a
mint + burn.  However, we were not checking for NFT updates before (we
were simply minting dupe assets with the same name and identifiers).
This commit adds the ability for the mint machine to sniff existing NFTs
(only for 1-of-1 quantity, uses blockfrost), and if so, perform an
automatic mint+burn combination.

NOTE: We refactor several parts of the code to use Javascript's native
await/async methodology to improve readability.

[ Documentation: confirm() box to display this understanding to the user ]
[ Testing: Manually overwrote token on mainnet ]