Whitelist: Implement send-to-self for vending-machines with whitelisting
This change uses the lucid APIs to collect all assets that exist
starting with a specified whitelist ID and then send them to the
selected wallet's address using Lucid.

NOTE: We make the choice not to explicitly collect UTxOs using lucid's
`collectFrom` API because we don't want to be in charge of ensuring the
payment balances.  Instead, the transaction builder will do that for us
from the entire wallet.

[ Documentation: None ]
[ Testing: rm -fr dist/ && npm run build; manual QA ]