Drop Payment: Enable exclusions for specific used WL assets
We want to allow the user to provide exclusions that they manage for
whitelist objects to ensure that the user is not sending UTxOs with
repeat objects multiple times.  However, we don't want to be
heavy-handed with prescribing a way to do it so we simply provide a
generic input array.

NOTE: The implementation in `drop-payment.html` is simply one example
that retrieves a set of used whitelist assets from CloudFlare pages.

[ Documentation: None ]
[ Testing: Manually see whitelist assets ]