Sweep: Add sniffer for error message returned by API
For some not totally clear reason there is an interplay between Lucid
and the wallet API that is causing submitted transactions to return a
string occassionally (this may have to do with the fact that the
specific error triggered was an experimental API `getCollateral` and
nothing to do with normal errors like a wallet disconnect).

So, we add a quick sniff test to see if we actually got a Tx Hash back
or some other error message.  If something else, we rethrow it.

[ Documentation: None ]
[ Testing: Manual QA ]