Vending Machine: Move custom-built Lucid fork code into our codebase
Previously, we were delegating to a custom function called `inputsOf`
that we attempted to encapsulate in Lucid but that was not successfully
merged (see PR https://github.com/spacebudz/lucid/pull/40).  instead, we
simply move that function to this class as `getInputsFor`and simplify
some of the error handlings to streamline the code.

[ Documentation: None, internal ]
[ Testing: E2E on preprod and mainnet ]