Mint Machine: Using message and input listeners to improve slot gen
Before, slots were only auto-generated when the actual slot was changed
or when "Mint Now" was clicked.  However, this is confusing for the user
who might want to write it down but didn't realize it changed.  So, we
add two listeners to improve the user experience and allow a more
seamless slot update.

[ Documentation: None, slot automatically displayed ]
[ Testing: Manual QA of expected behavior ]