Voting: Basic ballot creation prototype using Helios plus Lucid
This is a very basic on-chain minting policy that creates a proxy NFT
if-and-only-if the NFT being referenced was validly spent as an output
in the transaction.  What this enables us to do is to create a "ballot"
representing the NFTs under a specific policy that a user already owns
so they can drop it into a "ballot box" with their (vote) forthcoming
change.

NOTE: This uses the Helios programming language
(https://hyperion-bt.github.io/Helios-Book/The_Helios_Programming_Language.html)
as a fun exploration of on-chain minting policies directly from
JavaScript.  However, due to a bug in Aiken (see
https://github.com/Hyperion-BT/Helios/issues/4) we disable the native
untyped Plutus Core compilation (which forces Lucid to submit the
transaction to the Blockfrost evaluation backend).

NOTE: We update to Lucid 0.6.6 to ensure that any bugs we find are not
related to versioning updates.

Forthcoming commits need to investigate:
* Paying to a script that only specific oracles/vote counters can redeem
* UIs for selecting the desired vote
* Genericization of all the code herein to remove const/statics/etc.

[ Documentation: None ]
[ Testing: Manually created single NFT using on-chain policy ]