Sweep: Allow marketplace datum directly returned for JPG.Store
JPG.Store stores the datum for the transaction in the metadata of the
listing transaction, so that will need to be passed to the transaction
side.  For this change, we simply add a few available methods to the
`datumFor` function to support this.  All other logic will occur on the
browser extension side.

[ Documentation: None ]
[ Testing: None ]