Voting: Support large wallets and optimize Helios compilation
We anticipate there being large whales that have voting power of 100+
NFTs and want to mint 100+ ballots.  While we do not want to impact the
user experience, we need a way to "chunk" these so that the user can
cast votes incrementally.  This does not change the UI (total voting
power showed), it just adds Javascript confirmations of the chunked
voting behavior and checkpoints each vote.

NOTE: Current constraint is 9 due to Plutus limitations but we are
hoping for at least 30 soon.

NOTE: Compilation optimization flags provide a minimal memory
improvement (~1,800 units), but still worth putting in to reduce chain
fees at the very least.

[ Documentation: Adding JavaScript alerts to the workflow ]
[ Testing: Manual QA of various scenarios ]