Multi-Sig: Provide sample backend for Tangz Holiday mint
This is half of a system for the multi-signature mint that was
implemented in JavaScript under multisig-demo.html.

In this half, we stand up an AWS Chalice microservice (see
https://github.com/aws/chalice for documentation) that very simply:
1. Generates images based on user input
2. Uploads the images to NFT.storage when the user goes to mint
3. Validates and completes partially signed multi-sig mint txns

Note that the `pycardano` library is used for deserialization and object
access, as well as signing the transaction (once validated) using the
policy private key.

Because the only witness sent over the wire is the user's witness, we
also have to manually add back the NativeScript instance to the final
transaction object.

We then submit using blockfrost-python
(https://github.com/blockfrost/blockfrost-python).

[ Documentation: None ]
[ Testing: Ran through manually locally and on AWS with edge cases ]