Transactions: Prototype for web-based transaction scraper
This is a simple prototype to use Blockfrost to scrape historical
transactions for a specific stake key.  The UI is ugly.  Future
iterations should automatically categorize the transactions as one of 8
types:

1. List
2. Delist
3. Buy
4. Sell
5. Send-to-Self
6. Simple Pay
7. Simple Receive
8. Mint Asset

NOTE: This change requires a disconnection between Lucid and the wallet
connector so that you don't need to connect wallet for the transaction
scraping.

[ Documentation: None ]
[ Testing: Manual ]