Transactions: Change mechanism of token-gating to include "SUM" logic
The built-in token-gate map is simply an "OR" map so that any policy
holdings above a certain threshold are triggered.  Since this is a
collaboration with another project, we want to be sure that either
community can use it as long as the total holdings sum to a combined
threshold.  This change fixes that.

[ Documentation: None ]
[ Testing: Manual with combined holdings ]