WalletConnect: Update to cardano-dapp-js 1.0.8
This is to add support for WalletConnect (see the commit history at
cardano-dapp-js: https://github.com/thaddeusdiamond/cardano-dapp-js).

NOTE: This requires us to switch to a module-style export for the
`nft-toolkit.js` output otherwise there are syntactical errors in
webpack's output.  This will require manual migrations for all live
updates.

[ Testing: npm run build, static-server ]